# backend/authn/internal/adapters/inbound/httpapi/gen/api.models.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.7.1 DO NOT EDIT.
package api

import (
	openapi_types "github.com/oapi-codegen/runtime/types"
)

const (
	BearerAuthScopes bearerAuthContextKey = "BearerAuth.Scopes"
)

// Defines values for UserType.
const (
	UserTypeDEVICE         UserType = "DEVICE"
	UserTypeSERVICEACCOUNT UserType = "SERVICE_ACCOUNT"
	UserTypeUSER           UserType = "USER"
)

// Valid indicates whether the value is a known member of the UserType enum.
func (e UserType) Valid() bool {
	switch e {
	case UserTypeDEVICE:
		return true
	case UserTypeSERVICEACCOUNT:
		return true
	case UserTypeUSER:
		return true
	default:
		return false
	}
}

// Defines values for UserBaseType.
const (
	UserBaseTypeDEVICE         UserBaseType = "DEVICE"
	UserBaseTypeSERVICEACCOUNT UserBaseType = "SERVICE_ACCOUNT"
	UserBaseTypeUSER           UserBaseType = "USER"
)

// Valid indicates whether the value is a known member of the UserBaseType enum.
func (e UserBaseType) Valid() bool {
	switch e {
	case UserBaseTypeDEVICE:
		return true
	case UserBaseTypeSERVICEACCOUNT:
		return true
	case UserBaseTypeUSER:
		return true
	default:
		return false
	}
}

// Error defines model for Error.
type Error struct {
	// Details Additional error details
	Details *map[string]interface{} `json:"details,omitempty"`

	// Error Error code or type
	Error string `json:"error"`

	// Message Human-readable error message
	Message string `json:"message"`
}

// LoginRequest defines model for LoginRequest.
type LoginRequest struct {
	// Email User's email address
	Email openapi_types.Email `json:"email"`

	// Password User's password
	Password string `json:"password"`
}

// LoginResponse defines model for LoginResponse.
type LoginResponse struct {
	// AccessToken Opaque access token
	AccessToken string `json:"accessToken"`

	// ExpiresIn Token expiration time in seconds
	ExpiresIn int `json:"expiresIn"`

	// TokenType Token type
	TokenType string `json:"tokenType"`
	User      User   `json:"user"`
}

// MessageResponse defines model for MessageResponse.
type MessageResponse struct {
	// Message Response message
	Message string `json:"message"`
}

// PasswordResetConfirm defines model for PasswordResetConfirm.
type PasswordResetConfirm struct {
	// NewPassword New password (minimum 8 characters)
	NewPassword string `json:"newPassword"`

	// Token Password reset token received via email
	Token string `json:"token"`
}

// PasswordResetRequest defines model for PasswordResetRequest.
type PasswordResetRequest struct {
	// Email User's email address
	Email openapi_types.Email `json:"email"`
}

// ULID Universally Unique Lexicographically Sortable Identifier
type ULID = string

// User defines model for User.
type User struct {
	// Email User's unique email address (login identifier)
	Email openapi_types.Email `json:"email"`

	// FirstName User's first name
	FirstName *string `json:"firstName,omitempty"`

	// Id Universally Unique Lexicographically Sortable Identifier
	Id ULID `json:"id"`

	// LastName User's last name
	LastName *string `json:"lastName,omitempty"`

	// Locale User's locale
	Locale *string `json:"locale,omitempty"`

	// Timezone User's timezone
	Timezone *string `json:"timezone,omitempty"`

	// Type Principal type for extensibility
	Type UserType `json:"type"`
}

// UserType Principal type for extensibility
type UserType string

// UserBase defines model for UserBase.
type UserBase struct {
	// Email User's unique email address (login identifier)
	Email openapi_types.Email `json:"email"`

	// Id Universally Unique Lexicographically Sortable Identifier
	Id ULID `json:"id"`

	// Type Principal type for extensibility
	Type UserBaseType `json:"type"`
}

// UserBaseType Principal type for extensibility
type UserBaseType string

// UserCreateRequest defines model for UserCreateRequest.
type UserCreateRequest struct {
	// Email User's unique email address (login identifier)
	Email openapi_types.Email `json:"email"`

	// FirstName User's first name
	FirstName *string `json:"firstName,omitempty"`

	// LastName User's last name
	LastName *string `json:"lastName,omitempty"`

	// Locale User's locale
	Locale *string `json:"locale,omitempty"`

	// Password User's password (minimum 8 characters)
	Password string `json:"password"`

	// Timezone User's timezone
	Timezone *string `json:"timezone,omitempty"`
}

// UserCredentials defines model for UserCredentials.
type UserCredentials struct {
	// Email User's unique email address (login identifier)
	Email openapi_types.Email `json:"email"`

	// Password User's password (minimum 8 characters)
	Password string `json:"password"`
}

// UserProfile defines model for UserProfile.
type UserProfile struct {
	// FirstName User's first name
	FirstName *string `json:"firstName,omitempty"`

	// LastName User's last name
	LastName *string `json:"lastName,omitempty"`

	// Locale User's locale
	Locale *string `json:"locale,omitempty"`

	// Timezone User's timezone
	Timezone *string `json:"timezone,omitempty"`
}

// UserUpdateRequest defines model for UserUpdateRequest.
type UserUpdateRequest struct {
	// FirstName User's first name
	FirstName *string `json:"firstName,omitempty"`

	// LastName User's last name
	LastName *string `json:"lastName,omitempty"`

	// Locale User's locale
	Locale *string `json:"locale,omitempty"`

	// Timezone User's timezone
	Timezone *string `json:"timezone,omitempty"`
}

// ServiceUnavailable defines model for ServiceUnavailable.
type ServiceUnavailable = Error

// bearerAuthContextKey is the context key for BearerAuth security scheme
type bearerAuthContextKey string

// ConfirmUserRegistrationParams defines parameters for ConfirmUserRegistration.
type ConfirmUserRegistrationParams struct {
	// Token Confirmation token sent to the user's email
	Token string `form:"token" json:"token"`
}

// LoginUserJSONRequestBody defines body for LoginUser for application/json ContentType.
type LoginUserJSONRequestBody = LoginRequest

// ResetPasswordJSONRequestBody defines body for ResetPassword for application/json ContentType.
type ResetPasswordJSONRequestBody = PasswordResetConfirm

// RequestPasswordResetJSONRequestBody defines body for RequestPasswordReset for application/json ContentType.
type RequestPasswordResetJSONRequestBody = PasswordResetRequest

// RegisterUserJSONRequestBody defines body for RegisterUser for application/json ContentType.
type RegisterUserJSONRequestBody = UserCreateRequest

// UpdateUserProfileJSONRequestBody defines body for UpdateUserProfile for application/json ContentType.
type UpdateUserProfileJSONRequestBody = UserUpdateRequest