# backend/authn/internal/adapters/inbound/httpapi/openapi.yaml
openapi: 3.1.0
info:
  title: Beholder Authentication Service API
  description: |
    API for user authentication, profile management, and password operations using opaque tokens.
    Email is the unique primary identifier for all users.
  version: 0.0.1
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: https://api.example.com/v1
    description: Production server
  - url: https://staging-api.example.com/v1
    description: Staging server
  - url: http://localhost:8080/v1
    description: Local development server
...
paths:
...
  /authn/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: Creates a new user account with the provided credentials
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        '201':
          description: User successfully registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /authn/confirm:
    get:
      tags:
        - Authentication
      summary: Confirm user registration
      description: Confirms a user's email address using a confirmation token
      operationId: confirmUserRegistration
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
          description: Confirmation token sent to the user's email
      responses:
        '200':
          description: User successfully confirmed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '410':
          description: Token expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /authn/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticates a user and returns an opaque access token
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /authn/logout:
    post:
      tags:
        - Authentication
      summary: User logout
      description: Invalidates the current user's access token
      operationId: logoutUser
      security:
        - BearerAuth: [ ]
      responses:
        '204':
          description: Logout successful
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /authn/password/reset-request:
    post:
      tags:
        - Password
      summary: Request password reset
      description: Sends a password reset link to the user's email
      operationId: requestPasswordReset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
      responses:
        '202':
          description: Password reset email sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /authn/password/reset:
    post:
      tags:
        - Password
      summary: Reset password
      description: Resets the user's password using a valid reset token
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetConfirm'
      responses:
        '200':
          description: Password successfully reset
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Invalid request data or token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '410':
          description: Reset token expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /profile:
    get:
      tags:
        - Profile
      summary: Get user profile
      description: Retrieves the authenticated user's profile information
      operationId: getUserProfile
      security:
        - BearerAuth: [ ]
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'

    put:
      tags:
        - Profile
      summary: Update user profile
      description: Updates the authenticated user's profile information
      operationId: updateUserProfile
      security:
        - BearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /users/{id}:
    get:
      tags:
        - Users
      summary: Get user by ID
      description: Retrieves a user by their unique identifier
      operationId: getUserByID
      parameters:
        - name: id
          in: path
          required: true
          description: User's unique identifier
          schema:
            $ref: '#/components/schemas/ULID'
      responses:
        '200':
          description: User retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Opaque authentication token
  schemas:
...
    ULID:
      type: string
      pattern: '^[0-9A-HJKMNP-TV-Z]{26}$'
      description: Universally Unique Lexicographically Sortable Identifier
      example: 01H8R5X7X9Y2Z4W6V8T0S1Q2R
    UserBase:
      type: object
      required:
        - id
        - type
        - email
      properties:
        id:
          $ref: '#/components/schemas/ULID'
          description: The user id
        type:
          type: string
          enum:
            - USER
            - SERVICE_ACCOUNT
            - DEVICE
          description: Principal type for extensibility
        email:
          type: string
          format: email
          description: User's unique email address (login identifier)
          example: user@example.com
    UserCredentials:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User's unique email address (login identifier)
          example: user@example.com
        password:
          type: string
          format: password
          minLength: 8
          description: User's password (minimum 8 characters)
          example: SecurePass123!
    UserProfile:
      type: object
      properties:
        firstName:
          type: string
          description: User's first name
          example: John
        lastName:
          type: string
          description: User's last name
          example: Doe
        locale:
          type: string
          description: User's locale
          example: en-CA
        timezone:
          type: string
          description: User's timezone
          example: America/Toronto


    # =============================================================================
    # COMPOSED SCHEMAS - Built from atomic schemas
    # =============================================================================
    User:
      allOf:
        - $ref: '#/components/schemas/UserBase'
        - $ref: '#/components/schemas/UserProfile'
    UserCreateRequest:
      allOf:
        - $ref: '#/components/schemas/UserCredentials'
        - $ref: '#/components/schemas/UserProfile'
    UserUpdateRequest:
      type: object
      properties:
        firstName:
          type: string
          description: User's first name
          example: John
        lastName:
          type: string
          description: User's last name
          example: Doe
        locale:
          type: string
          description: User's locale
          example: en-CA
        timezone:
          type: string
          description: User's timezone
          example: America/Toronto

    # =============================================================================
    # REQUEST/RESPONSE SCHEMAS
    # =============================================================================
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User's email address
          example: user@example.com
        password:
          type: string
          format: password
          description: User's password
          example: SecurePass123!
    LoginResponse:
      type: object
      required:
        - accessToken
        - tokenType
        - expiresIn
        - user
      properties:
        accessToken:
          type: string
          description: Opaque access token
          example: 5f9a7b2c4e3d1a6f8b9c0e2d4f6a8c1b
        tokenType:
          type: string
          description: Token type
          example: Bearer
        expiresIn:
          type: integer
          description: Token expiration time in seconds
          example: 3600
        user:
          $ref: '#/components/schemas/User'
    PasswordResetRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          description: User's email address
          example: user@example.com
    PasswordResetConfirm:
      type: object
      required:
        - token
        - newPassword
      properties:
        token:
          type: string
          description: Password reset token received via email
          example: a1b2c3d4e5f6g7h8i9j0
        newPassword:
          type: string
          format: password
          minLength: 8
          description: New password (minimum 8 characters)
          example: NewSecurePass456!
    MessageResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: Response message
          example: Operation completed successfully
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error code or type
          example: INVALID_CREDENTIALS
        message:
          type: string
          description: Human-readable error message
          example: The provided credentials are invalid
        details:
          type: object
          additionalProperties: true
          description: Additional error details
  responses:
    ServiceUnavailable:
      description: Service Temporarily Unavailable or Degraded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Suggests a duration in seconds to wait before retrying
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'